Run scripts directly instead of building
The scripts in this repo are primarily intended for developers to run;
they are not deployed anywhere, and in particular do not need to be run
outside of npm or in a context without dev dependencies installed.

Switch from the two-step build & run approach to running the TypeScript
files directly via ts-node. Update the GitHub CI step to run tsc
directly, and move the dev tsconfig over the main tsconfig.