Fetch token using client credentials
In order to interact with the PDC API, we need a token to authenticate
with. Add a method to fetch the token, and add a command to confirm that
token retrieval works.

Fetching a token requires several bits of configuration: the location of
the OIDC authority, the client ID, and the client secret. Add parameters
to collect this information, and replace the placeholder configuration
environment and config file keys (which demonstrated how to pass
configuration data to yargs) with the new OIDC configuration keys.

Decoding the token is solely for debugging support, but it saves us
having to paste tokens into https://jwt.io or the like.