Update Candid data for all organizations in PDC
Add the last command for storing Candid data in the PDC data platform
provider cache: update-all. Fetch the list of applications from the PDC,
find their EINs, call Candid with each EIN, and upload the response from
Candid to the PDC.

Happily, as this command does not include any positional arguments and
so does not need to use the yargs function builder API, the type
definitions are much better.