Upload a single Candid result to the PDC
Given an EIN, look up an organization in the Candid Premier API, and
upload it to the PDC data platform provider cache.

Unfortunately, the type inference around yargs commands is not quite
powerful enough. I spent some time fighting the type checker and
couldn't get it to do the right thing around reused `Options`; it just
isn't able to figure them out, or at least I haven't been able to figure
out how to help it do so correctly. It works better when the `builder`
is a literal instead of a function, but positional arguments cannot be
specified via the literal syntax, and the `check` validator is also only
available via the function builder. So, sadly, back to sprinkling `as
string` around.