Use `csv-parse`'s sync API throughout the scripts
Without this change there would be two CSV reading libraries in use
and two scripts would continue to have race conditions. We do not need
the sophistication of a stream API because we have yet to see any CSVs
larger than a few MiB. This change uses the technique already present in
the `postProposalVersions.ts` script in the older two scripts.

Because there would be repeated boilerplate otherwise, a new `csv.ts`
has the commonly used type and type assertion guard.

Issue #25 Race between row and EOF events for forms and proposals