Handle async errors
Yargs handles commands that throw synchronous errors differently from
commands that return rejected promises. The example code introduced in
commit 863379dc09beb52a14cf4685acae175fe8f906d3 threw a synchronous
error, but the error returned by axios is via a rejected Promise.

Configure yargs to not include its standard error handling behavior, so
that our error handler will fire, and update the example error command
to be async so as to exercise that behavior.

Thanks to @bickelj for catching this!

https://github.com/yargs/yargs/blob/main/docs/api.md#failfn--boolean